requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.66
botocore==1.35.66
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
databases==0.9.0
distro==1.9.0
fastapi==0.115.5
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
jmespath==1.0.1
openai==1.55.0
psycopg2-binary==2.9.10
pydantic==2.10.0
pydantic_core==2.27.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
